% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{anti_plularity}
\alias{anti_plularity}
\title{anti plularity,each voter votes against a single candidate, and the candidate with the fewest votes against wins}
\usage{
anti_plularity(scores)
}
\arguments{
\item{scores}{voters scores}
}
\value{
winner
}
\description{
anti plularity,each voter votes against a single candidate, and the candidate with the fewest votes against wins
}
