% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{approbal}
\alias{approbal}
\title{Approbal vote}
\usage{
approbal(situation, mode = "seuil")
}
\arguments{
\item{situation}{voters preferences}

\item{mode}{n_approbation mode}
}
\value{
winner
}
\description{
Approbal vote
}
