% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{condorcet_looser}
\alias{condorcet_looser}
\title{Condorcet Looser}
\usage{
condorcet_looser(preference_matrix)
}
\arguments{
\item{preference_matrix}{pref matrix}
}
\value{
looser
}
\description{
Condorcet Looser
}
