% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{condorcet_winner}
\alias{condorcet_winner}
\title{Condorcet Winner}
\usage{
condorcet_winner(preference_matrix)
}
\arguments{
\item{preference_matrix}{pref matrix}
}
\value{
winner
}
\description{
Condorcet Winner
}
