% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{condorcetV2}
\alias{condorcetV2}
\title{CondorcetV2}
\usage{
condorcetV2(preference_matrix)
}
\arguments{
\item{preference_matrix}{voters preferences}
}
\value{
winner, can be NULL
}
\description{
CondorcetV2
}
