% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usefull_fonctions.R
\name{draw_test}
\alias{draw_test}
\title{draw_test}
\usage{
draw_test(pref_matrix, remaining_candidates)
}
\arguments{
\item{pref_matrix}{voters preferences}
}
\value{
draw - false/true
}
\description{
draw_test
}
