% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{find_winner}
\alias{find_winner}
\title{Succesif elimination test}
\usage{
find_winner(pref_matrix)
}
\arguments{
\item{pref_matrix}{voters preferences}
}
\value{
remaining_candidates
}
\description{
Succesif elimination test
}
