% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{majority_jugement}
\alias{majority_jugement}
\title{Majority Jugement (Balinski) medianne}
\usage{
majority_jugement(situation)
}
\arguments{
\item{situation}{voters preferences}
}
\value{
winner
}
\description{
Majority Jugement (Balinski) medianne
}
