% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usefull_fonctions.R
\name{make_duel_matrix}
\alias{make_duel_matrix}
\title{Make duel matrix}
\usage{
make_duel_matrix(preference_matrix)
}
\arguments{
\item{preference_matrix}{voters preferences}
}
\value{
duel_matrix
}
\description{
Make duel matrix
}
