% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{minimax}
\alias{minimax}
\title{Minimax procedure}
\usage{
minimax(preference_matrix)
}
\arguments{
\item{preference_matrix}{voters preferences}
}
\value{
winner, can be NULL
}
\description{
Minimax procedure
}
