% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{nanson}
\alias{nanson}
\title{Nanson method}
\usage{
nanson(pref_matrix, first_it = TRUE)
}
\arguments{
\item{pref_matrix}{voters preferences}
}
\value{
remaining_candidates
}
\description{
Nanson method
}
