% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usefull_fonctions.R
\name{points_to_preferences}
\alias{points_to_preferences}
\title{points_to_preferences}
\usage{
points_to_preferences(points_matrix, eliminated_candidates)
}
\arguments{
\item{points_matrix}{voters preferences}
}
\value{
preferences_matrix
}
\description{
points_to_preferences
}
