% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usefull_fonctions.R
\name{preferences_to_points}
\alias{preferences_to_points}
\title{Preferences_to_points}
\usage{
preferences_to_points(preferences)
}
\arguments{
\item{preferences}{voters preferences}
}
\value{
ranks
}
\description{
Preferences_to_points
}
