% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usefull_fonctions.R
\name{reallocate_points}
\alias{reallocate_points}
\title{reallocate_points}
\usage{
reallocate_points(points_matrix, eliminated_candidates)
}
\arguments{
\item{points_matrix}{voters preferences}

\item{eliminated_candidates}{eliminated_candidates}
}
\value{
preferences_matrix
}
\description{
reallocate_points
}
