% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usefull_fonctions.R
\name{reallocate_preferences}
\alias{reallocate_preferences}
\title{reallocate_preferences}
\usage{
reallocate_preferences(pref_matrix, eliminated_candidates)
}
\arguments{
\item{pref_matrix}{voters preferences}

\item{eliminated_candidates}{eliminated candidates}
}
\value{
matrice_sans_zero
}
\description{
reallocate_preferences
}
