% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usefull_fonctions.R
\name{reallocate_preferencesV3}
\alias{reallocate_preferencesV3}
\title{reallocate_preferencesV3}
\usage{
reallocate_preferencesV3(pref_matrix, eliminated_candidates)
}
\arguments{
\item{pref_matrix}{voters preferences}

\item{eliminated_candidate}{eliminated index}
}
\value{
reallocated_prefs
}
\description{
reallocate_preferencesV3
}
