% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usefull_fonctions.R
\name{rearrange_points}
\alias{rearrange_points}
\title{Rearrange borda points}
\usage{
rearrange_points(points_matrix, eliminated_candidates)
}
\arguments{
\item{points_matrix}{borda_points}

\item{eliminated_candidates}{eliminated_candidates}
}
\value{
preferences_matrix
}
\description{
Rearrange borda points
}
