% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usefull_fonctions.R
\name{scores_to_borda_points}
\alias{scores_to_borda_points}
\title{scores_to_borda_points}
\usage{
scores_to_borda_points(scores)
}
\arguments{
\item{scores}{voters preferences}
}
\value{
points
}
\description{
scores_to_borda_points
}
