% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usefull_fonctions.R
\name{scores_to_preferences}
\alias{scores_to_preferences}
\title{Scores_to_preferences}
\usage{
scores_to_preferences(scores)
}
\arguments{
\item{scores}{voters scores}
}
\value{
ranks
}
\description{
Scores_to_preferences
}
