% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{star}
\alias{star}
\title{star, 2 meilleures moyenne et on regarde le JM sur les 2}
\usage{
star(scores)
}
\arguments{
\item{scores}{voters scores}
}
\value{
winner
}
\description{
star, 2 meilleures moyenne et on regarde le JM sur les 2
}
