% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{successif_elimination}
\alias{successif_elimination}
\title{Succesif elimination test}
\usage{
successif_elimination(pref_matrix)
}
\arguments{
\item{pref_matrix}{voters preferences}
}
\value{
remaining_candidates
}
\description{
Succesif elimination test
}
