% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{uninominal}
\alias{uninominal}
\title{Uninominal vote}
\usage{
uninominal(situation, n_round = 1)
}
\arguments{
\item{situation}{voters preferences}

\item{n_round}{int}
}
\value{
winner_idx
}
\description{
Uninominal vote
}
