% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{uninominal_vote}
\alias{uninominal_vote}
\title{Uninominal vote}
\usage{
uninominal_vote(situation, n_round = 1)
}
\arguments{
\item{situation}{voters preferences}

\item{n_round}{int}
}
\value{
winner_idx
}
\description{
Uninominal vote
}
