% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{vote_elimination}
\alias{vote_elimination}
\title{Succesif elimination}
\usage{
vote_elimination(situation)
}
\arguments{
\item{situation}{voters preferences}
}
\value{
remaining
}
\description{
Succesif elimination
}
