% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{approval}
\alias{approval}
\title{Approval vote}
\usage{
approval(scores, mode = "autre")
}
\arguments{
\item{scores}{voters scores}

\item{mode}{n_approbation mode}
}
\value{
winner
}
\description{
Approval vote
}
