% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{nanson}
\alias{nanson}
\title{Nanson method}
\usage{
nanson(scores, first_it = TRUE)
}
\arguments{
\item{scores}{voters scores}

\item{first_it}{first iteration}
}
\value{
winner
}
\description{
Nanson method
}
