% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{range_voting}
\alias{range_voting}
\title{Range voting (vote à la moyenne)}
\usage{
range_voting(situation)
}
\arguments{
\item{situation}{voters preferences}
}
\value{
winner
}
\description{
Range voting (vote à la moyenne)
}
