% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usefull_fonctions.R
\name{reallocate_preferences}
\alias{reallocate_preferences}
\title{reallocate_preferences}
\usage{
reallocate_preferences(pref_matrix, eliminated_candidate)
}
\arguments{
\item{pref_matrix}{voters preferences}

\item{eliminated_candidate}{eliminated index}
}
\value{
mat
}
\description{
reallocate_preferences
}
