% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{successif_elimination}
\alias{successif_elimination}
\title{Succesif elimination}
\usage{
successif_elimination(pref_matrix, first_it = TRUE)
}
\arguments{
\item{pref_matrix}{voters preferences}

\item{first_it}{first iteration}
}
\value{
remaining_candidates
}
\description{
Succesif elimination
}
