% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{uninominal}
\alias{uninominal}
\title{Uninominal vote}
\usage{
uninominal(scores, n_round = 1)
}
\arguments{
\item{scores}{voters scores}

\item{n_round}{int}
}
\value{
winner
}
\description{
Uninominal vote
}
