% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{uninominal_vote}
\alias{uninominal_vote}
\title{Uninominal vote}
\usage{
uninominal_vote(situation)
}
\arguments{
\item{situation}{voters preferences}
}
\value{
winner_idx
}
\description{
Uninominal vote
}
